Move post app content into forum as it's more relevant
