+++ VeteranTalk/settings.py
@@ -48,7 +48,6 @@ INSTALLED_APPS = [
     "about",
     "forums",
     "home",
-    "posts",
     "users",
 ]
