+++ forums/admin.py
@@ -1,3 +1,9 @@
 from django.contrib import admin
+from .models import Post
 
-# Register your models here.
+
+class PostAdmin(admin.ModelAdmin):
+    pass
+
+
+admin.site.register(Post, PostAdmin)
