+++ forums/urls.py
@@ -1,7 +1,4 @@
 from django.urls import path
 from . import views
 
-
-urlpatterns = [
-    path("", views.ForumView.as_view(), name="forum"),
-]
+urlpatterns = [path("", views.PostListView.as_view(), name="post_list")]
