--- posts/admin.py
@@ -1,9 +1,0 @@
-from django.contrib import admin
-from .models import Post
-
-
-class PostAdmin(admin.ModelAdmin):
-    pass
-
-
-admin.site.register(Post, PostAdmin)
